refactor(chart): single chart-root shell (PlotSurface); PlotChart + Plot.Root reuse it
PlotChart (Plot.svelte) and Plot.Root independently reimplemented the entire root —
PlotState+context, ResizeObserver→width, animate gate, DefinePatterns, svg+canvas shell,
zoom, selection — which is why fixes (patterns, responsive width, mode/animate defaults)
kept landing on one side only and diverging.

Extract PlotSurface.svelte as the ONE shell; both roots now render it:
- Plot.Root = thin wrapper (build config from props -> PlotSurface).
- PlotChart = PlotSurface + its batteries (grid/axes/geoms/trend/highlight as canvas
  children; legend/tooltip/a11y-table via an overlay snippet so they stay INSIDE the
  plot-state context) + title.
- Selection sync, zoom, responsive width, animate gate, patterns now live once.

A future root fix lands everywhere. Live-verified every chart type (20 sections render,
all stretch to fill, selection/legend/a11y-table/trend/highlight/grid all work, console
clean); full suite 1302, svelte-check + lint clean.
